% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrystallinitY.R
\docType{data}
\name{PLS_MR}
\alias{PLS_MR}
\title{%CI and Raman spectra of MCC batches acquired using a MR probe}
\format{

}
\usage{
PLS_MR
}
\description{
%CI and Raman spectra of MCC batches acquired using a MR probe
}
\keyword{datasets}
