% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrystallinitY.R
\docType{data}
\name{PLS_PhAT}
\alias{PLS_PhAT}
\title{%CI and Raman spectra of MCC batches acquired using a PhAT probe}
\format{

}
\usage{
PLS_PhAT
}
\description{
%CI and Raman spectra of MCC batches acquired using a PhAT probe
}
\keyword{datasets}
